Improved it a bit
`lock ()` no longer returns the file descriptor. Instead it's closed at
the end of the function if creation is a success. Don't have to call
`close (2)` at every `unlock ()` call which means not having to have the
file descriptor a global variable (or worse, a function parameter). Less
refactoring too :)